fix(audit): code-audit of PR #491 scroll appearance/behavior
Multi-pass /code-audit of the scroll appearance/behavior feature.

Correctness / consistency:
- scrollhandler: minimized scroll windows no longer keep a hidden title
  bar — onWindowMinimizedChanged restores it on minimize and re-hides on
  restore; refreshDecorations/updateHideTitleBarsSetting skip minimized
  windows so the state is not clobbered. Previously a minimized window
  showed chrome-less (no title bar, no border) in the overview.
- scrollhandler: notifyWindowAdded decorates just the new window
  (setWindowBorderless + updateWindowBorder) instead of rebuilding the
  whole strip via refreshDecorations() — avoids O(N) border teardown per
  open. onWindowMinimizedChanged likewise switched to per-window.
- scrollhandler: setWindowBorderless verifies window identity before
  setNoBorder(false), so findWindowById's appId fuzzy fallback cannot
  restore an unrelated same-app window's title bar.
- scrollhandler: notifyWindowFocused keeps the focus-follows-mouse dedup
  key current on non-FFM focus changes; added a null guard to
  notifyWindowAdded for parity with sibling methods.
- borders.cpp: dropped dead `m_scrollHandler &&` null-guards — both
  handlers are constructed unconditionally and outlive the effect.
- settingsschema: the two scroll border QColor keys gained validColorOr
  validators (every other QColor key has one); key order matched to
  tilingAppearanceColorsGroup.

File-size rule (CLAUDE.md <800 lines), both crossed by this feature:
- daemon_bringup.cpp (826): slotSettingsChanged + loadSettingAsync +
  loadCachedSettings extracted to new TU plasmazoneseffect/settings_load.cpp.
- scrollhandler.cpp (804): the border-decoration functions extracted to
  new TU scrollhandler/decoration.cpp.

Tests:
- test_settings_core: added round-trip coverage for the 9 new scroll
  appearance/behavior settings and the use-system-border-colors path.

Build green, 191/191 tests pass.
