+++ kwin-effect/CMakeLists.txt
@@ -45,6 +45,7 @@ set(plasmazones_effect_SRCS
     plasmazoneseffect/window_lifecycle.cpp
     plasmazoneseffect/daemon_bringup.cpp
     plasmazoneseffect/daemon_apply.cpp
+    plasmazoneseffect/settings_load.cpp
     plasmazoneseffect/drag_snap.cpp
     plasmazoneseffect/paint_pipeline.cpp
     plasmazoneseffect/shader_transitions.cpp
@@ -62,6 +63,7 @@ set(plasmazones_effect_SRCS
     screenchangehandler.cpp
     screenchangehandler.h
     scrollhandler.cpp
+    scrollhandler/decoration.cpp
     scrollhandler.h
     snapassisthandler.cpp
     snapassisthandler.h
